# Add ZEND_DIV spec handlers