https://github.com/niklas-heer/speed-comparison

For this synthetic (and probably meaningless) benchmark we get a significant speedup (10000000 rounds).

> hyperfine -- "../php-src/sapi/cli/php -d zend_extension=$(pwd)/../php-src/modules/opcache.so -d opcache.enable_cli=1 test.php"

```log
Before:
Benchmark 1: ../php-src/sapi/cli/php -d zend_extension=/home/ilutov/Developer/symfony_demo/../php-src/modules/opcache.so -d opcache.enable_cli=1 test.php
  Time (mean ± σ):     820.1 ms ±  12.1 ms    [User: 813.9 ms, System: 3.9 ms]
  Range (min … max):   809.1 ms … 849.8 ms    10 runs

After:
Benchmark 1: ../php-src/sapi/cli/php -d zend_extension=/home/ilutov/Developer/symfony_demo/../php-src/modules/opcache.so -d opcache.enable_cli=1 test.php
  Time (mean ± σ):     775.7 ms ±   7.0 ms    [User: 768.7 ms, System: 4.8 ms]
  Range (min … max):   765.8 ms … 786.0 ms    10 runs
```

I didn't see a big different for Symfony Demo, so I'm not sure if this is worth adding, although it might not hurt.