@iluuu1994 2022-12-17
@dstogov Thank you for your assessment. Let's close this then. I don't want to risk performance degradation in real applications to improve a benchmark. At least this PR helped me to understand how to benchmark apps more accurately (CGI + warmup).